fix ts-config
