+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.11",
   "description": "A quest for consistency",
   "main": "./dist/index.js",
+  "module": "./dist/index.es.js",
   "types": "./dist/index.d.ts",
   "sideEffects": false,
   "files": [
@@ -13,8 +14,11 @@
   },
   "scripts": {
     "prebuild": "rm -rf dist",
-    "build": "tsc",
-    "test": "nyc ava",
+    "postbuild": "yarn run types",
+    "build": "rollup -c",
+    "dev": "rollup -c -w",
+    "types": "yarn dts-bundle-generator -o ./dist/index.d.ts ./dist/gen-types/index.d.ts",
+    "test": "TS_NODE_COMPILER_OPTIONS='{\"module\":\"commonjs\"}' nyc ava",
     "lint": "yarn run eslint --ext .ts src",
     "coverage": "nyc report --reporter=lcov > coverage.lcov && npx codecov",
     "prepare": "yarn run -s build"
@@ -36,10 +40,15 @@
     "eslint-plugin-jsx-a11y": "^6.2.1",
     "eslint-plugin-react": "^7.12.4",
     "esm": "^3.2.22",
+    "dts-bundle-generator": "^2.1.0",
     "husky": "^1.3.1",
     "lint-staged": "^8.1.5",
     "nyc": "^13.3.0",
     "react": "^16.8.6",
+    "rollup": "^1.6.0",
+    "rollup-plugin-node-resolve": "^4.0.1",
+    "rollup-plugin-terser": "^4.0.4",
+    "rollup-plugin-typescript2": "^0.20.1",
     "ts-node": "^8.0.3",
     "typescript": "^3.4.3"
   },
@@ -56,14 +65,18 @@
   },
   "nyc": {
     "all": true,
+    "include": [
+      "src/*.ts"
+    ],
+    "exclude": [
+      "dist",
+      "test"
+    ],
     "extension": [
       ".ts"
     ],
-    "exclude": [
-      "**/*.d.ts",
-      "node_modules/**",
-      "coverage/**",
-      "test/**"
+    "require": [
+      "ts-node/register"
     ],
     "reporter": [
       "lcov",
