+++ tsconfig.json
@@ -7,7 +7,7 @@
     "module": "esnext",
     "moduleResolution": "node",
     "declaration": true,
-    "declarationDir": "./dist",
+    "declarationDir": "./dist/gen-types",
     "allowSyntheticDefaultImports": true,
     "esModuleInterop": true,
     "noImplicitAny": false,
